feat(frontend): add AI assistant dialog
